+++ centreon/packages/ui/src/components/Modal/ModalHeader.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement, ReactNode } from 'react';
 
-import { DialogTitle as MuiDialogTitle } from '@mui/material';
+import { DialogTitleProps, DialogTitle as MuiDialogTitle } from '@mui/material';
 
 import { useStyles } from './Modal.styles';
 
@@ -8,12 +8,17 @@ export type ModalHeaderProps = {
   children?: ReactNode;
 };
 
-const ModalHeader = ({ children }: ModalHeaderProps): ReactElement => {
+const ModalHeader = ({
+  children,
+  ...rest
+}: ModalHeaderProps & DialogTitleProps): ReactElement => {
   const { classes } = useStyles();
 
   return (
     <div className={classes.modalHeader}>
-      <MuiDialogTitle color="primary">{children}</MuiDialogTitle>
+      <MuiDialogTitle color="primary" {...rest}>
+        {children}
+      </MuiDialogTitle>
     </div>
   );
 };
